atomic-ops: split ix86 and x86_64
Some immediate instructions require a size suffix (addq, subq, andq)
which differs between 32 and 64 bit operands and which cannot be
determined by the assembler. In addition, immediate operands on x86_64
require the "e" constraint instead of "i" because it is not possible
to pass values larger than 32 bits. Thus better split the code between
the two arches.

While doing so, the unused atomic_add64 and atomic_inc64 functions
were removed since their original purpose was to implement the "long"
variants of the operations while not applicable to 32-bit.