lock.c: remove exponential back-off from all polled locks
There are a number of places where we're only reading the lock, waiting
for the owner to release it. It's counter-efficient to apply an exponential
back-off there because as soon as the cache line is updated and broadcasted
we'll see the change. Instead perform a fast poll loop. This provides an
important performance boost, about 50% faster on 4 and 8 threads for 75%
read scenario.

The new performance for 1,2,3,4 and 8 threads looks like this :

                          Threads
R/W ratio      1        2        3        4         8
128  (50%)  1943269  3666384  3888049  4298329   4138242
192  (75%)  1939312  3800860  5668970  6682301   6882356
224  (87%)  1929024  3842483  5712690  7423952  10964982
240  (93%)  1925681  3849139  5719224  7564344  12106615
248  (96%)  1933687  3849139  5768713  7668760  12468907
252  (98%)  1930327  3857305  5753776  7651158  12666325
254  (99%)  1936495  3871492  5757089  7654086  12714639
256 (100%)  1925310  3856561  5763725  7686444  12861818