doc: refresh the doc and mention the write-atomic operations
Some combinations were wrong (eg: W doesn't get M anymore). Also mention
the rules for owner and others for each lock.

The "modify" or "frozen" lock is better called "seek lock". Indeed, the
purpose is to hold it while seeking for the place to modify. This explains
well why it's save to read in parallel and why it doesn't make sense to
allow two concurrent threads to seek for a place since one of them may
modify what the other will find.