lock: reimplement S lock acquisition
In order to reduce the pressure on other threads and limit cache line
bouncing, we now first check if the lock is free before trying to grab
it. This improved the treelock test performance by 10% when using 4
threads with 50% read ratio. It will also help with the implementation
of the atomic write locks later.

The lock was reimplemented by simply looping on the pl_try_s() function,
thus reducing code duplication. It was verified that the same code was
emitted.

The functions have been marked static instead of static inline so that
we don't force the compiler to inline them when it doesn't make sense.