lock: reimplement W lock acquisition
In order to reduce the pressure on other threads and limit cache line
bouncing, we now first check if the lock is free before trying to grab
it. This improved the treelock test performance by 12% when using 4
threads with 50% read ratio in mode 4 (wr-only locks).

The lock was reimplemented by simply looping on the pl_try_w() function,
thus reducing code duplication.

The functions have been marked static instead of static inline so that
we don't force the compiler to inline them when it doesn't make sense.