lock: implement shared atomic access (A)
The A lock allows multiple compatible users to perform atomic writes
but protects R, S, W and X locks against such modifications. This
lock is made only of a W bit. It is slightly stronger than the X lock
since it has to support W > 1. The X lock was enhanced to detect this
situation. However the S and W locks remain stronger than the A lock
since the S bit is a promise of ability to complete an operation.

It is technically possible to atomically upgrade a read-only (R) lock
to A if the requester accepts the risk of failure. This will allow
certain "pick()" operations to run in parallel even during the seek
phase. It can also allow a cleaning task or garbage collector to run
over a linked list and delete some elements once in a while without
having to lock the whole list read-only.