lock: implement R->S transition
This one is opportunistic. It is a good alternative to the S-locked
lookups for workloads having a high read ratio, typically a cache.
Measurements have shown that it's faster than S-locked lookups as
soon as the read-to-write ratio is higher than 2:1. However the
performance can become disastrous when a lot of writes are involved
because the caller will have to perform the failed operation again.

A cache may decide to maintain a cache hit ratio to decide what strategy
to use for its lookups.