plock: allow to immediately grab the W lock
Till now we didn't implement immediate access to the W lock as
it either required two operations (take_s + stow) or would have
caused an integer overflow by trying to grab W and S at the same
time. The X lock was implemented for this purpose, but it suffered
from a fairness issue (it could be starved by R).

In fact it turns out that the integer overflow is not an issue because
we have two bits for the S counter and W+S have a multiple of 4 bits.
By adding W|S, we in fact add the integer value 5 to the W:S word, which
cannot reach zero. Once 13108 writers are contented on 32 bits or
858993460 on 64 bits, we loop back to value 4 (W=1, S=0) and this
protects against all other locks but the atomic lock. And since the
W lock now holds the R lock at the same time, this last one doesn't
overflow and W+S+R are indeed safe against the A lock.

This is a nice improvement, in terms of API (as the X lock can now be
removed), and in terms of performance. A quick test on treelock replacing
the X lock by an immediate W lock for 2 threads at 50% read, 50% write
show a speed increase from 2226194 loops/s to 2338922 (+5%).