plock: implement type-agnostic locking
We now support both 32- and 64- bit locks, solely based on the size of the
argument passed to the locking functions. For this, the functions all had
to be replaced by macros, generally written as compound integer expressions.

The resulting code is very slightly smaller than the previous one relying
on static inlines, apparently due to a compiler scheduling change in one
"if" block involving builtin_expect(), that had to be reordered to fit
with no return statement.