tests: add lrubench to measure the locks performance
The benchmark models a very simple LRU cache. The idea is to measure
different locking mechanism for read-intensive workloads, like caches. A
typical LRU cache will be used to cache the result of some expensive
computations. Here as "expensive computation", we turn an integer into its
text representation. And to make it a bit heavier than a single snprintf()
call, we iterate over it 100 times so that a successful lookup is cheaper
than a computation. The storage is made of a simple hash table. Random keys
are emitted, and for each of them, a lookup is performed in the cache. If
the key is found, the associated string is used. Otherwise the string is
computed and the result stored into the cache. After an insertion, the cache
is trimmed to ensure it's never larger than the configured size. The result
is then checked using atoi() to verify that it matches (which mimmicks the
use of the looked up data). A lock is added so that it's possible to use
this cache from multiple threads. For two seconds, several threads run this
workload in parallel, looking up random values in this shared cache. The
number of operations per thread is collected and aggregated at the end, then
reported.

The default settings use a 32-head hash table for the cache, a size limit
to 100 keys per entry on average (hence 3200 keys), and a random key space
of 101%, causing 1% miss and 99% hit.

The observed performance with 8 threads running on a quad-core 8-threads
core i7-6700k is :

 mode  loops/s  ns/loop notes
  1    2923835   342    pthread spinlock lock for everything
  2    4034342   247    pthread rwlock : R lock for lookup, W for insertion
  3    3189377   313    plock W lock for everything
  4    3190205   313    plock S lock for everything
  5   18292612    54    plock R lock for lookup, W for insertion
  6   18682354    53    plock R lock for lookup, S->W for insertion
  7   18602670    53    plock R lock for lookup, R->S->W for insertion