plock: divide reads by two on uncontended case
Till now, in order to limit the risk of disturbance between threads,
we first perform a read to check for compatibility between the lock's
state and the new desired state before doing the operation. But in the
case of low contention, this means one read cycle, then another
read-modify-write cycle.

For most operations we don't need to to this and can tentatively perform
the operation, and watch the value in case of conflict.

This patch does this for the S, W and A locks. The contention is handled
by a call to pl_wait_unlock_<type>(). It's important to use exponential
back-off here as we're more aggressive on the first attempt, so we want
to relax after the first failure.r

Counter-intuitively, the R lock did not benefit from this and instead
slowed things down. The reason is that the W lock waits for all R locks
to leave after being grabbed. But if R lock requesters (generally
numerous) are too aggressive, the W lock takes longer to see all R
locks disappear. Thus the change was not performed. In theory the same
would happen between the S and W locks, except that this contention is
not frequent.

This patch only touches the pl_take_*() functions, and not the pl_try_*()
since these ones are expected to be used without disturbing. This results
in a significant boost on certain cases, such as when the W or S locks
alone are used (+20-25% for 8 threads).