plock: implement a set of functions to wait for a lock release
Right now all the code used to wait for a lock release is inlined
and a bit heavy due to the involved loop. These new functions
pl_wait_unlock_long() and pl_wait_unlock_int() respectively wait
for the long or int locks to release the bits passed in the designated
mask. They're marked static and not inline, so that they can be shared
between most lock locations, saving some code size.

And since they're implemented as non-inlined functions, it's cheap to
implement an exponential backoff. Here 2 is used as the basis of the
exponentiation, bounded to 1..2^8-1, and CPU pause calls are performed
during this time. This considerably helps the owner thread to complete
its job by avoiding to force another core to publish its local changes.
The gain is even more visible in SMT environments where it allows the
second thread running on the same core to run faster. Gains of up to
~46% were observed when running on S locks using the LRU cache bench in
mode 4 with 90% hit ratio with 24 threads :
  - before : rate(lps): 666785, access(ns): 1499, misses=129755
  - after  : rate(lps): 976178, access(ns): 1024, misses=188603

At this stage, the functions are not used yet, this will be done in the
next commit.