tests: lrubench: add an argument to set the cost of a miss.
It defaults to 100 but may be increased or decreased to test for various
workloads. The minimum cost is 1 (it's the number of iterations performed
aroudn snprintf() to write the string into the cache item), and using 0
will still result in at least one iteration being done.

It's visible that the differences fade away when the cost increases, and
that mode 7 gets slightly better than mode 6 when the cost drops below 10.