WIP: plock: implement hardware lock elision for W locks
This partially implements HLE for W locks (and only them). The performance
gain is impressive :
  - *4 for W-only lock
  - *1.5 for R+S>W locks

WARNING! the try* functions, the 32-bit versions etc were not handled.

Apparently there is a requirement to return the lock to its original
state on release, which makes this not much compatible with doing this
for other locks like R or maybe S.

It is possible that this represents a nice solution for ebtrees.