plock: implement rtoj, jtoc, rtoc, and ctoa
These ones are the upgradable multi-writer locks. Just the minimal set
of functions was implemented for now.

Exponential back-off is implemented for rtoj and rtoc, as reaching these
can take quite some time and test shows that performance benefits from
doing this. However, it's not applied to ctoa because work done during
this step is supposed to be extremely fast (compare and swap) so it does
not make much sense to add the cost of a back-off herei, which is also
confirmed by benchmarks.