tests: implement a minimal test for the multiple writer model
This is a copy of testlock.c to testmw.c where the write access
was replaced with an upgrade to J/C/A (in fact R->C->A as tests
show that saving one memory read slightly improves performance).
The lookup cost is set to 300 and the modification to 30.

First tests show only a minimal degradation as thread count grows
when the lookup cost is small (<100 loops), but a decent scalability
when the lookup cost increases (1000) :

At lookup_cost=1000 : 4 threads are 2.68 times faster than 1 :
  $ ./testmw -t 1 -r 0
  threads: 1 loops: 20000128 time(ms): 25338 rate(lps): 789333
  $ ./testmw -t 2 -r 0
  threads: 2 loops: 20000128 time(ms): 15247 rate(lps): 1311741
  $ ./testmw -t 3 -r 0
  threads: 3 loops: 20000128 time(ms): 11465 rate(lps): 1744450
  $ ./testmw -t 4 -r 0
  threads: 4 loops: 20000128 time(ms): 9449 rate(lps): 2116639

At lookup_cost=300 : 4 threads are 1.71 times faster than 1 :
  $ ./testmw -t 1 -r 0
  threads: 1 loops: 20000128 time(ms): 8372 rate(lps): 2388930
  $ ./testmw -t 2 -r 0
  threads: 2 loops: 20000128 time(ms): 6407 rate(lps): 3121605
  $ ./testmw -t 3 -r 0
  threads: 3 loops: 20000128 time(ms): 5478 rate(lps): 3650990
  $ ./testmw -t 4 -r 0
  threads: 4 loops: 20000128 time(ms): 4896 rate(lps): 4084993