plock: add a few missing "register" modifiers on some local variables
We really want that the elements manipulated in the loops are assigned to
registers only because if the compiler decides to use a stack variable
instead due to register pressure, this can have a disastrous performance
impact caused by the implicit barriers created by memory clobbering,
causing reloads to be performed all along. Usually it has no effect
except in some rare situations, so it's not expected to have hit use
case.