plock: implement pl_take_j() and pl_drop_j()
This one is a special case, only one thread may succeed on this operation,
which is a way to guarantee an exclusive access to the shared resource in
the R->J->C->A model. This operation will first wait for any writer (J,C,A)
to leave and will prevent any other R,J,C from coming next. Concurrent A
will wait for it to terminate.