plock: add two new functions to wait on a lock : pl_wait_new_{int,long}
These functions wait for a change on a lock's value while applying an
exponential back-off, and return the new value. They are written to be
both small and efficient for the fast path (lock taken first time) and
the asm code was checked for this.

The purpose is to call them from a while loop like this :

    curr = pl_deref_long(lock);
    while (curr & particular_mask)
        curr = pl_wait_new_long(lock, curr);

For situations where it's expected that the value changes often and we want
to leave only when a specific mask is released, pl_wait_unlock_{int,long}
may remain slightly more optimal though.