examples: add an example providing an alternative to pthread's rwlock
During tests involving HAProxy and OpenSSL on a large scale system, it
was found that the performance didn't scale at all beyond a few tens of
cores and that all the time (up to 80-85%) is spent in the kernel dealing
with spinlocks just for futex_wake(). The reason is that since version
1.1.0, OpenSSL's libcrypto ignores the provided locking mechanism and
uses pthread's rwlocks instead, while for some code paths (e.g. async
engines) this results in a huge amount of syscalls on systems facing a
bit of contention.

This patch provides an alternative by redefining the relevant pthread
rwlocks from the low-overhead version of the progressive rw locks. This
way there will be no more syscalls in case of contention, and CPU will
be burnt in userland. Doing this saves massive amounts of CPU, where
the locks only take 12-15% vs 80% before, which allows SSL to work much
faster on large thread counts (e.g. 24 or more).

The tryrdlock and trywrlock variants have been implemented using a CAS
since their goal is only to succeed on no contention and never to wait.
The pthread_rwlock API is complete except that the timed versions of
the rdlock and wrlock do not wait and simply fall back to trylock
versions.

This version does *not* implement exponential back-off though other
experiments were done with and without it, showing inconsistent
results depending on platforms and programs: sometimes the extra
latency caused by the exp-bo overrides the gains provided by the
very low latency locking. Such an implementation could be added in
the future so that the user can choose which one to pick.

The code simply needs to be built with:

   cc -O2 -c pth_rwl.c -pthread

and the resulting .o will have to be linked into the executable.
Alternatively the .c file may directly be copied into the target
project if its license permits it (it generally should).