lorw: low-overhead algorithm for non-upgradable RW locks
This R/W lock tries to minimize the number of memory accesses per
participant because on highly contended systems, it's what always
kills the performance. It also focuses on being able to figure from
the lock's value whether it's locked for exclusive access or not, so
that it can work with a common unlock() operation which only involves
one extra load.

This one focuses on the number of operations in the fast path for locks
that usually run under low contention, while still taking the effort of
keeping the contended path fast. And it is compatible with pthread_rwlock
(though no try operation are provided even though they're easy to
implement).

The lock is slightly slower than the regular upgradable locks on both
small and large systems, but it becomes faster when the need of the
common unlock operation appears because it saves the previous one from
having to perform an extra store on success.

Read and write operations are mutually exclusive in that the write
will obviously wait for reads to leave, but if a write detects a
conflicting read, it will also prevent them from coming before being
elected. This tends to preserve a bit of ordering of operations and
to help reads being grouped.

It was tuned based on measured times. It was indeed found that a CAS
seems to reserve the cache line for a few cycles after a failure (which
is not always desired but sometimes helps), that a BTS is always more
expensive than a CAS (maybe due to the implied bit shift?), an XADD is
roughly the same cost as a BTS, and that a simple OR/AND is faster than
XADD, BTS and CAS.

An even better approach seems possible using split words (CAS on the
low word to avoid the failures caused by the unknown values at the top),
and stores to release the lock.

It gives this:
    reader:
       success: 1 XADD
       failure: 1 + 2N XADD
       unlock:  1 sub
    writer:
       success: 1 RD + 1 CAS
       failure: 1 RD + 1 CAS + 1 XADD + N BTS
       unlock:  1 sub

This results in very low latency on both small and large machines, typically
15M/s on core i7 and still 7M/s on a much larger CPU count (48 threads). Two
new test numbers were added to lrubench (10 and 11).