plock: let the exponential back-off call sched_yield() when available
For cases where threads compete on the same CPU (e.g. process reload),
being able to give the CPU back would be a bonus. That's what we now do
using sched_yield when _POSIX_PRIORITY_SCHEDULING is defined before
including that file. This one will be set on compatible platforms if
the locks are included after unistd.h. It's interesting to note that
even when calling sched_yield() it's important to keep a large number
of pauses. It seems that the reason is that if there's nothing to
reschedule, the kernel can return quickly enough for the loop to
spin too fast. By only removing 1/2 to 3/4 of the loop's count we
continue to spend time in the pause loop.