plock: improve the default exp-bo algo
The exponential back-off isn't very efficient because it limits itself
to a large value that saturates and tends to create some resonance
effect when there are many CPUs. In an ideal exponential back-off
mechanism, some random ought to be involved, but calculating randoms
costs a lot in this context, even xorshift is expensive as it requires
to store a temporary value. In addition, the risk of collision with
many CPUs following discrete time is high, and using a power of two
tends to make all of them execute pause loops at the same time.

A better approach was experimentally found to work well both on low
core count machines and high core count ones. It consists in using a
non-power of two grows factor coupled with a large mask. This creates
mostly non-repetitive sequences when the values wrap around the mask
instead of saturating, and significantly reduces the resonance effects.
For example, the current algorithm:

	m = ((m * 3 / 2) | 2) & 0x7fff

was experimentally found to provide very good latencies and produces
this sequence showing exponential increases and wrap arounds creating
sawtooths:

  2 3 6 11 18 27 42 63 94 143 214 323 486 731 1098 1647 2470 3707
  5562 8343 12514 18771 28158 9471 14206 21311 31966 15183 22774
  1395 2094 3143 4714 7071 10606 15911 23866 3031 4546 6819 10230
  15347 23022 1767 2650 3975 5962 8943 13414 20123 30186 12511
  18766 28151 9458 14187 21282 31923 15118 22679 1250 1875 2814
  4223 6334 9503 14254 21383 32074 15343 23014 1755 2634 3951 ...

Another growth factor close to e (~2.7, 11/4 used here) was attempted,
but didn't prove to be as good (in general anything getting away from
1.5 reduces the efficiency).

Also, with such large values it is tempting to pack multiple pauses
at once. It appears that a single pl_cpu_relax() call followed by
jumps does not pause long, but that when chaining a few it gets much
better. Decounting them by packs of 4 and switching to a single at
once only when < 4 initially semeed better in certain circumstances
but this also degrades other areas significantly, so for now we'll
avoid it.

Finally since we know the last value seen, let's return it as some
callers could benefit from using it.