plock: improve the exponential back-off for different CPU types
The latest changes improved the lock latency on Xeon CPUs but resulted
in a significant slowdown on an EPYC CPU, where the pause instruction
seems to last extremely long. This patch changes the back-off so that
we stop using pause when only a few hundred loops are expected. However
this required to decrement the slower steps by a larger value so now
we decrement pause by 10 and empty loops by 1.

In the end the performance was recovered on EPYC and even slightly
improved on Xeon.