atomic: use an ARMv8 instruction barrier for the pause instruction
As suggested by @AGSaidi in haproxy issue #958, on ARMv8 its convenient
to use an "isb" instruction in pl_cpu_relax() to improve fairness.
Without it I've met a few watchdog conditions on valid locks with 16
threads, indicating that some threads couldn't manage to get it in 2
seconds. It never happened again with it. In addition, the performance
increased by slightly more than 5% thanks to the reduced contention.

This was first merged into haproxy as commit 1e237d037 ("MINOR: plock:
use an ARMv8 instruction barrier for the pause instruction").