plock: make the rtos upgrade faster
With many threads it's obvious that performing the xadd before even
reading is beneficial. Among the few places that were still doing a
preliminary read is pl_try_rtos() which is used to attempt a lock
upgrade. Generally this is used in low-to-moderate contention code,
so it's supposed to succeed most often. Better pay the rollback price
in a few rare occasions than pay the price of two accesses to the same
line, with the first one not being that relevant in the end. Some tests
have shown up to a 5% request rate gain on haproxy.