atomic: use C11 atomics when available
This can be cleaner over the long term, especially when we want to
add different memory orders.

However the approach here is still not perfect because the arch-specific
definitions still have precedence, despite the asm() statements offering
less flexibility to the compiler in certain cases. Practically speaking,
in the current situation the C11 atomics will still not be used for x86
due to all operations already being defined there.

Some operations that do not exist as-is in the C11 atomics also are not
implemented yet (e.g. bts/btr etc) but it's not a problem, as the fallback
is still present. A next step should bring the ability to select which
definitions have precedence on a case-by-case basis.