atomic: guard all generic definitions to make it easier to provide specific ones
The pl_xxx definitions are grouped by architecture but that's not
necessarily convenient, especially when other definitions are expected
to be added because, because this requires to add all implementations
at once. Let's instead guard the generic ones and move them out of the
global ifdef to ease redefinition. Now they can safely be used as a
fallback, and the arch-specific ones only preset the preferred ones
when there are some.