atomic: prefer the C11 definitions in general
The C11 __atomic_* tend to have more intimacy with the compiler than
the asm() operations because the compiler knows how not to emit certain
instructions if needed. However we don't necessarily have all of them
available, and it's still possible that some asm definitions remain
better for certain corner cases.

This patch sets two level of definitions, the preferred and the fallback
ones. The preferred ones have their final names while the fallback ones
are prefixed with "_". The arch-specific code sets the fallback ones
unless it knows better, while the C11 code sets the preferred ones. After
the C11 definitions we check for missing preferred ones and set them from
the fallback ones if needed. Doing so allows us to have all definitions
and to always be able to support the best implementation available.