atomic: add more _noret operations
Actually some of our arithmetic operations as implemented in asm are
ambiguous:
 - pl_inc/pl_dec return a value
 - pl_inc_noret / pl_dec_noret do not
 - pl_add/and/or/sub/xor do not either
 - pl_xadd does

Worse, the generic versions based on the __sync API actually *do* return
the post-op value as the API might suggest, adding more confusion since
some code might build on some archs and not on x86.

The reason for some not returning a value above is because when the
result is not needed, direct memory operations are simpler on x86
(no CAS needed). But the API is clearly too confusing, so let's add
new _noret variants to all operations so that add/and/or/sub/xor now
also behave like their inc/dec counterpart.

For now what is done is that these _noret operations are defined as a
pure copy of the other ones, but cast as void, to avoid mistakes in the
future (and this will remain the fallback solution).