atomic: handle modern compilers that support returning flags
A few operations (inc/dec/bts/btr) were still having an instruction
dedicated to assigning a result to a register based on a flag, so
that the register could be tested. Reasonably modern compilers that
set __GCC_ASM_FLAG_OUTPUTS__ are able to return a flag in a variable
that the compiler can optimize away for jumps or loops. Let's use
that when available.