atomic: add the fetch-and-<op> operations (pl_ld<op>)
The pl_ld<op> operations perform a load (fetch) before the operation and
return the previous value. This is the same as the x86 "xswap". They're
supported in C11 atomic as "atomic_fetch_and_<op>()". Now they will be
consistent. These were added with the seq_cst memory order semantic.
pl_xswap was not renamed and remains as-is, but the new ones will be
extended with memory ordering constraints.