plock: use a release ordering when dropping the lock
However downgrades continue to use a full barrier. Normally that should
not be needed since downgrades are atomic so no other thread may have
changed the memory. But there's still the case of going from W to R and
at the same time another might upgrade from R to S. In theory it's not
allowed to perform writes during this phase, and if it does it could do
them after the lock was dropped anyway, so that should be safe; But let's
keep changes simple for now.