plock: use EBO when waiting for readers to leave in take_w() and stow()
Once the writer is known for being the only one, it currently waits for
exising readers to leave. However till now it was done in a busy loop,
which can possibly make upgrades and write locks expensive. Let's use
the EBO-based wait functions for this instead.

The same could be done in pl_try_w(), pl_try_rtow() and the transitions
to the various (less used) A/J/C states.