plock: switch to CAS + XADD for pl_take_r()
The non-contented stick-table tests on haproxy show that the read lock
is expensive to grab because of the first read. And this makes sense:
  - there are plenty of places where nobody holds it most of the time,
    but reading before incrementing it requires two memory accesses ;

  - on read-mostly contention, reading it and incrementing it can
    result in one extra round trip of the cache line, forcing another
    core to drop it, while a 3rd one can grab it again between our
    read and the attempt to increment it.

Some tests were performed by incrementing it first, and showed that
the non-contented stick-tables tests raised from 4.2 to 5.6M RPS by
just doing this! However the write contention test decreased since
an extra write needs to be corrected.

A better solution consists in solving the non-contended case using a
CAS to switch 0 to 1. In case of success there's nothing else to do,
the lock is acquired. However in case of contention, this will fail
and we can immediately take action:
  - it's only a read contention, we can instantly retry an XADD which
    has great chances of instant success on such locks.

  - it's a write contention, then we wait for the writer to leave
    without adding any extra write cycle, allowing the writer to
    leave faster.

With this method, we've preserved the performance under write
contention (~750k/s on AMD, ~610k/s on ARM) and improved the nominal
case with read-only contention from 4.2 to 5.6M on ARM and 2.0 to 2.05M
on AMD. Note that this change is particularly dependent on the previous
EBO update that reduces the number of local loops.