tests: rwrace: detect races between read and write locks
This detects bugs such as the one fixed by commit a68f743, and also
reports the number of loops performed by the reader and the writers,
and will stop after 10s. Enabling EBO inlining allows to see where
time is spent, and disabling EBO allows to compare the performance.

This test was contributed by Remi Tricot-Le Breton and was slightly
adjusted to be usable as-is here.